TASK-INPUT now handles the different times.
As a result of the previous change, compilation of the
DEFINE-OPERATION form requires TASK-INPUT objects to exist. Thus any
DEFINE-TASK-INPUT forms need to be evaluated at compile time.

The resulting forms produced by DEFINE-OPERATION now contain
TASK-INPUT instances, thus a MAKE-LOAD-FORM is required so that these
forms can be compiled.